requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.0

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

pydantic==2.5.0

python-jose[cryptography]==3.3.0

black==23.12.1